Moved classes from animated/testing to /testing in imagepipeline
