+++ imagepipeline/src/test/java/com/facebook/imagepipeline/animated/base/AnimatedDrawableTest.java
@@ -20,13 +20,13 @@ import android.graphics.drawable.Drawable;
 
 import com.facebook.common.references.CloseableReference;
 import com.facebook.common.references.ResourceReleaser;
-import com.facebook.common.testing.FakeClock;
+import com.facebook.imagepipeline.testing.FakeClock;
 import org.robolectric.RobolectricTestRunner;
 import com.facebook.imagepipeline.animated.testing.MyShadowBitmap;
 import com.facebook.imagepipeline.animated.testing.MyShadowCanvas;
 import com.facebook.imagepipeline.animated.testing.TestAnimatedDrawableBackend;
 import com.facebook.imagepipeline.animated.impl.AnimatedDrawableDiagnosticsNoop;
-import com.facebook.imagepipeline.animated.testing.TestScheduledExecutorService;
+import com.facebook.imagepipeline.testing.TestScheduledExecutorService;
 
 import com.nineoldandroids.animation.ValueAnimator;
 import org.junit.Before;
