+++ imagepipeline/src/test/java/com/facebook/imagepipeline/testing/TrivialPooledByteBuffer.java
@@ -1,8 +1,13 @@
-// Copyright 2004-present Facebook. All Rights Reserved.
+/*
+ * Copyright (c) 2015-present, Facebook, Inc.
+ * All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
+ */
 
 package com.facebook.imagepipeline.testing;
-
-import java.io.InputStream;
 
 import com.facebook.imagepipeline.memory.PooledByteBuffer;
